requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
faiss-cpu
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
python-dotenv
tiktoken